rename and remove extra props
